[EasyApiPlatform] Refactor Symfony bundle configuration (#1209)
